tidy up, console.log written for errors
