+++ intern/pages/api/dbLink/attendance.ts
@@ -9,7 +9,6 @@ import {
 
 const mainURL = mainRoute();
 
-// // Function to grab participant attendance data from Zoom API and add to our database
 export default async function attendanceZoomToDb() {
   // Grab zoom data of everyone, latest meeting ID
   const ZoomparticipantsJSON = await fetch(
@@ -19,53 +18,9 @@ export default async function attendanceZoomToDb() {
     `${mainURL}${getRoute}getListBootcampers`
   );
 
-  // //variable/ reponse clean up
+  //variable/ reponse clean up
   const ZoomparticipantsClean = await ZoomparticipantsJSON.text();
   const zoomParticipants = JSON.parse(ZoomparticipantsClean);
-
-  // This is an example of how data should look like after we run zoomParticipants
-  //   // const parsedData = {
-  //   //   success: true,
-  //   //   participants: [
-  //   //     {
-  //   //       registrant_id: "_ic4PfRWS1mCf5GsC3Zhcg",
-  //   //       name: "(Kit) Wing-Kit Leung",
-  //   //       user_email: "wing_kitleung@hotmail.com",
-  //   //       day: "2023-12-14T",
-  //   //       duration: 182,
-  //   //       join_leave: [
-  //   //         {
-  //   //           join: "2023-12-14T10:11:12Z",
-  //   //           leave: "2023-12-14T10:11:15Z",
-  //   //         },
-  //   //         {
-  //   //           join: "2023-12-14T10:11:15Z",
-  //   //           leave: "2023-12-14T10:12:17Z",
-  //   //         },
-  //   //         {
-  //   //           join: "2023-12-14T10:12:17Z",
-  //   //           leave: "2023-12-14T10:12:22Z",
-  //   //         },
-  //   //         {
-  //   //           join: "2023-12-14T10:12:23Z",
-  //   //           leave: "2023-12-14T10:12:32Z",
-  //   //         },
-  //   //         {
-  //   //           join: "2023-12-14T10:12:32Z",
-  //   //           leave: "2023-12-14T10:12:41Z",
-  //   //         },
-  //   //         {
-  //   //           join: "2023-12-14T10:12:41Z",
-  //   //           leave: "2023-12-14T10:12:45Z",
-  //   //         },
-  //   //         {
-  //   //           join: "2023-12-14T10:12:46Z",
-  //   //           leave: "2023-12-14T10:14:16Z",
-  //   //         },
-  //   //       ],
-  //   //     },
-  //   //   ],
-  //   // };
 
   const allParticipantsClean = await AllParticipantsJSON.text();
   const allParticipants = JSON.parse(allParticipantsClean);
@@ -75,6 +30,7 @@ export default async function attendanceZoomToDb() {
   console.log(zoomParticipants);
   console.log("AllParticipants");
   console.log(allParticipants);
+
   // //start loop
 
   for (let i = 0; i < zoomParticipants.participants.length; i++) {
@@ -82,59 +38,133 @@ export default async function attendanceZoomToDb() {
       `running loop ${i + 1} of ${zoomParticipants.participants.length}`
     );
     const currentZoomID = zoomParticipants.participants[i].registrant_id;
+    // const currentZoomID = "12345678";
     const duration = zoomParticipants.participants[i].duration;
+    let iszoomIdMatched = false;
+
+    // debug logger
+    // console.log(`currentZoomID: ${currentZoomID}`);
+    // console.log(`duration: ${duration}`);
+
     // If zoomid then do following:
     //see if zoom id exists
+
     for (let j = 0; j < allParticipants.data.length; j++) {
       console.log(
         `running inner loop ${j + 1} of ${allParticipants.data.length}`
       );
-      if (!currentZoomID.includes(allParticipants.data[j].zoomid)) {
+
+      //Goes through database looking for currentZoomID. If match, then will make variable iszoomIdMatched true
+      if (currentZoomID === allParticipants.data[j].zoomid) {
         try {
           console.log(
-            `zoom participants : ${zoomParticipants.participants[i].name}`
+            `zoom ids match : ${currentZoomID} ${allParticipants.data[j].zoomid}`
           );
-          // if doesnt then add bootcamper to database
+          iszoomIdMatched = true;
+
+          // log if it errors
+        } catch (error) {
+          console.log("Error:", error);
+        }
+
+        //patch section
+        const newAttendanceHours = duration / 3600;
+        const newTotalAttendancehours =
+          allParticipants.data[j].total_attendance_hours + newAttendanceHours;
+        const newDays =
+          newAttendanceHours >= 7
+            ? (allParticipants.data[j].total_days_attended += 1)
+            : allParticipants.data[j].total_days_attended;
+        const newMissingStreak =
+          newAttendanceHours < 7
+            ? (allParticipants.data[j].missing_streak += 1)
+            : (allParticipants.data[j].missing_streak = 0);
+        console.log(
+          `new attendance hour: ${newAttendanceHours}, new total attendance : ${newTotalAttendancehours}, amount days attended:  ${newDays}, the missing streak: ${newMissingStreak}`
+        );
+        try {
+          console.log("alive at 85");
+
+          // updates entry with zoomAPI data
           const response = await fetch(
-            `${mainURL}${postRoute}postBootcamperAttendance?zoomId=${currentZoomID}`,
+            `${mainURL}${patchRoute}registerBootcamperAttendance?zoomId=${currentZoomID}`,
             {
               // set header
-              method: "POST",
+              method: "PATCH",
               headers: {
                 "Content-Type": "application/json",
               },
-              // send name in the body
-              body: JSON.stringify(zoomParticipants.participants[i].name),
+              // send new patch data to body
+              body: JSON.stringify({
+                todays_attendance_hours: newAttendanceHours,
+                total_attendance_hours: newTotalAttendancehours,
+                total_days_attended: newDays,
+                missing_streak: newMissingStreak,
+              }),
             }
           );
           // log that it works
-          // const result = await response.json();
+          console.log("alive 105");
+          // const result = await response;
           const resultJson = await response.text();
           const cleanResult = JSON.parse(resultJson);
-
           console.log("Success:", cleanResult);
           // log if it errors
         } catch (error) {
           console.log("Error:", error);
         }
       }
+    }
+
+    // If iszoomIdMatched is still false, i.e. currentZoomID isnt in the database
+    if (!iszoomIdMatched) {
+      // if zoomID isnt in database, then add bootcamper to database
+      console.log(`${currentZoomID} did not match any in database`);
+
+      try {
+        const response = await fetch(
+          `${mainURL}${postRoute}postBootcamperAttendance?zoomId='${currentZoomID}'`,
+          {
+            // set header
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            // send name in the body
+            body: JSON.stringify({
+              name: zoomParticipants.participants[i].name,
+            }),
+          }
+        );
+        console.log(`currentZoomID: ${currentZoomID}`);
+        console.log(`name: ${zoomParticipants.participants[i].name} `);
+
+        // log that it works
+        // const result = await response.json();
+        const resultJson = await response.text();
+        // console.log("response");
+        // console.log(response);
+        // console.log("resultJson");
+        // console.log(resultJson);
+        const cleanResult = JSON.parse(resultJson);
+        console.log("Success:", cleanResult);
+
+        // log if it errors
+      } catch (error) {
+        console.log("Error:", error);
+      }
+
+      // then patch current data:
       //patch section
-      const newAttendanceHours = duration / 3600;
-      const newTotalAttendancehours =
-        allParticipants.data[j].total_attendance_hours + newAttendanceHours;
-      const newDays =
-        newAttendanceHours >= 7
-          ? (allParticipants.data[j].total_days_attended += 1)
-          : allParticipants.data[j].total_days_attended;
-      const newMissingStreak =
-        newAttendanceHours < 7
-          ? (allParticipants.data[j].missing_streak += 1)
-          : (allParticipants.data[j].missing_streak = 0);
+      const newAttendanceHours = Number((duration / 3600).toFixed(1));
+      const newTotalAttendancehours = newAttendanceHours;
+      const newDays = newAttendanceHours >= 7 ? 1 : 0;
+      const newMissingStreak = newAttendanceHours < 7 ? 1 : 0;
       console.log(
         `new attendance hour: ${newAttendanceHours}, new total attendance : ${newTotalAttendancehours}, amount days attended:  ${newDays}, the missing streak: ${newMissingStreak}`
       );
       try {
-        console.log("alive at 117");
+        console.log("alive at 162");
         // if doesnt then add bootcamper to database
         const response = await fetch(
           `${mainURL}${patchRoute}registerBootcamperAttendance?zoomId=${currentZoomID}`,
@@ -154,7 +184,7 @@ export default async function attendanceZoomToDb() {
           }
         );
         // log that it works
-        console.log("alive 133");
+        console.log("alive 182");
         // const result = await response;
         const resultJson = await response.text();
         const cleanResult = JSON.parse(resultJson);
@@ -163,34 +193,10 @@ export default async function attendanceZoomToDb() {
       } catch (error) {
         console.log("Error:", error);
       }
+
+      let iszoomIdMatched = false;
     }
   }
-
-  //end loop if
-  //then do patch
-  // If no zoomid then insert new row
-
-  // Format information into this shape:
-  // Raw data from zoom:
-  // todays_attendance_hours = COALESCE ($1, todays_attendance_hours),
-  // WHERE zoomId = $5
-
-  // Convert duration (seconds) to hours.
-
-  // Get data from database
-
-  // Calculated from DB + Zoom
-  // total_attendance_hours = COALESCE ($2, total_attendance_hours),
-  // total_days_attended = COALESCE ($3, total_days_attended),
-  // missing_streak = COALESCE ($4, missing_streak)
-
-  // Create loop to patch by zoomid.
-
-  // console.log(getBootcampers);
-
-  // pool.end();
 }
-
-// //http://localhost:3000/api/dbLink/attendance
 
 attendanceZoomToDb();
